Data saved in TableSeeder

+++ database/migrations/2018_04_21_083837_create_subscriptions_table.php
@@ -15,7 +15,7 @@ class CreateSubscriptionsTable extends Migration
     {
         Schema::create('subscriptions', function ($table) {
             $table->increments('id');
-            $table->unsignedInteger('user_id');
+            $table->bigInteger('user_id');
             $table->string('name');
             $table->string('stripe_id');
             $table->string('stripe_plan');
